+++ extensions/easydict/src/releaseVersion/releaseNote.ts
@@ -2,53 +2,34 @@
  * @author: tisfeng
  * @createTime: 2022-07-03 22:10
  * @lastEditor: tisfeng
- * @lastEditTime: 2022-10-13 12:23
+ * @lastEditTime: 2022-10-26 22:05
  * @fileName: releaseNote.ts
  *
  * Copyright (c) 2022 by tisfeng, All Rights Reserved.
  */
 
 export const releaseNote = `
-## [v2.5.0] - 2022-10-12
-
-本次更新主要支持了**现代汉语词典**，方便查询中文词语的拼音和释义了。
-
-### ✨ 新功能
-
-- 新增支持有道现代汉语词典。
-
-![yi-1665582552](https://raw.githubusercontent.com/tisfeng/ImageBed/main/uPic/yi-1665582552.png)
-
-![ai-1665582591](https://raw.githubusercontent.com/tisfeng/ImageBed/main/uPic/ai-1665582591.png)
-
-![ai-1665582690](https://raw.githubusercontent.com/tisfeng/ImageBed/main/uPic/ai-1665582690.png)
-
+## [v2.6.1] - 2022-10-26
 
 ### 💎 优化
 
-- 优化了动作列表，现如果安装了 Eudic，会优先显示 'Open in Eudic' 动作，否则显示 'Copy Text' 动作。可以在设置中更改显示顺序。
-- 修改 Linguee 默认支持系统代理。
-- 修改所有列表都支持显示更多动作。
+- 优化代码。
+- 更新 README。
 
 ### 🐞 修复
 
-- 修复了许多已知问题。
+- 修复了最新 Raycast 1.42.0 版本中，arguments 为空会导致扩展崩溃的问题。
 
 #### 如果觉得这个扩展还不错，给个 [Star](https://github.com/tisfeng/Raycast-Easydict) ⭐️ 支持一下吧 (^-^)
 
 ---
 
-✨ New
+### 💎 Improvement
 
-- Added support for Youdao Modern Chinese Dictionary.
+- Optimized code.
+- Updated README.
 
-💎  Improvement
+### 🐞 Fixes
 
-- Improved the action list, now if Eudic is installed, the 'Open in Eudic' action will be shown first, otherwise the 'Copy Text' action will be shown. You can change the display order in the settings.
-- Changed Linguee to support system proxy by default.
-- Changed all lists to support 'Show More Details'.
-
-🐞 Fixes
-
-- Fixed many known issues.
+- Fixed the problem that the latest Raycast 1.42.0 version will cause the extension to crash when arguments is empty.
 `;