+++ extensions/easydict/src/releaseVersion/versionInfo.ts
@@ -28,10 +28,10 @@ export class Easydict {
 
   // new version info
   // * NOTE: this is new version info, don't use it directly. Use getCurrentStoredVersionInfo() instead.
-  version = "2.5.0";
-  buildNumber = 15;
-  versionDate = "2022-10-12";
-  isNeedPrompt = true;
+  version = "2.6.1";
+  buildNumber = 17;
+  versionDate = "2022-10-26";
+  isNeedPrompt = false;
   hasPrompted = false; // * always default false, only show once, then should be set to true.
   releaseMarkdown = releaseNote;