+++ extensions/toggl-track/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Toggl Track Changelog
 
- ## [New Feature] - 2022-05-23
+## [New Feature] - 2022-10-20
+
+- Add a manual refresh action to command's main window.
+
+## [New Feature] - 2022-05-23
 
 - Add billing support to the Toggl Track extension. Added the ability to toggle whether a new task is billable or not, as well as indicators to entries. The billable toggle only shows up if the project selected is billable (as determined by the track API)